refactor: move definition and initiallization to lib/referenceMarker/referenceMarkerContext.ts
