+++ src/components/Conversation/Editors/Viewpoints/FactListCard.tsx
@@ -12,7 +12,7 @@ import { MagnifyingGlassIcon, PlusIcon } from "@heroicons/react/24/outline";
 import { getPaginatedIssueFactsBySize } from "@/lib/requests/issues/getIssueFacts";
 import EditableViewpointReference from "@/components/Conversation/Editors/Viewpoints/EditableViewpointReference";
 import FactCreationModal from "@/components/Conversation/Facts/FactCreationModal";
-import { ReferenceMarkerContext } from "@/components/ReferenceMarker/ReferenceMarkerProvider";
+import { ReferenceMarkerContext } from "@/lib/referenceMarker/referenceMarkerContext";
 
 import type { Fact } from "@/types/conversations.types";
